syntax: reorganize number literals
This commit completely reorganizes the syntax highlighting for number
literals. The hareNumbers highlight group has been removed, as it is
unnecessary. Additionally, the hareFloat group is now linked to Float
rather than Number to better match the semantics of the default
highlight groups.

The syntax detection for floating-point numbers occurs later in the file
so as to prevent digits before and after '.' from being recognized as
two separate numbers.

Signed-off-by: Amelia Clarke <me@rsaihe.dev>