ftplugin: conform to Vim style
This commit makes a few minor adjustments to the Hare filetype plugin to
conform to the style used by Vim itself. Specifically, the preferred
indentation and textwidth settings for Hare are now gated behind
'g:hare_recommended_style' but are still set by default. Additionally,
the required 'b:undo_ftplugin' has been added.

Besides minor adjustments to formatting and the addition of 'comments'
and 'suffixesadd', '/' has been added to 'formatoptions', meaning that
comments are only continued by 'o' and 'O' if the comment is at the
beginning of the line. When the comment is at the end of a statement,
comments will not be continued:

	defer free(buf); // This comment will not be continued.

Signed-off-by: Amelia Clarke <me@rsaihe.dev>