Fix license formatting
This follows the formatting used in the official Vim license, hard
wrapping instead of having each paragraph on a single line.
Additionally, all instances of the word "Vim" in the license text that
were actually referring to this project have been replaced with
"hare.vim", to match the intended use of the license.

Signed-off-by: Amelia Clarke <selene@perilune.dev>